LPD-72034 Remove signer override configuration from S3Store
# breaking

## What modules/apps/portal-store/portal-store-s3/src/main/java/com/liferay/portal/store/s3/S3Store.java

Removed usage of 'signerOverride' from S3Store

## Why

Signers that were supported in AWS SDK v1 are no longer supported in v2. The
default and recommended signer type to use is now SigV4 (i.e. AwsV4HttpSigner).
AWS considers overriding the signer as an advanced configuration that should
only be done if necessary. It is technically possible to bring back the
deprecated signers, but they would need to be implemented from scratch.

----